Fix some races in testing

+++ nodes/stampzilla-server/logic/scheduler_test.go
@@ -9,8 +9,8 @@ import (
 func TestSchedulerAddTask(t *testing.T) {
 
 	scheduler := NewScheduler()
-	actionRunCount := 0
-	actionCancelCount := 0
+	actionRunCount := int64(0)
+	actionCancelCount := int64(0)
 	action := NewRuleActionStub(&actionRunCount, &actionCancelCount, t)
 
 	//Add first task.
@@ -36,8 +36,8 @@ func TestSchedulerAddTask(t *testing.T) {
 func TestSchedulerRemoveTask(t *testing.T) {
 
 	scheduler := NewScheduler()
-	actionRunCount := 0
-	actionCancelCount := 0
+	actionRunCount := int64(0)
+	actionCancelCount := int64(0)
 	action := NewRuleActionStub(&actionRunCount, &actionCancelCount, t)
 
 	//Add first task.
@@ -96,8 +96,8 @@ func TestSchedulersaveToFile(t *testing.T) {
 	scheduler := NewScheduler()
 	scheduler.ActionService = NewActions()
 
-	actionRunCount := 0
-	actionCancelCount := 0
+	actionRunCount := int64(0)
+	actionCancelCount := int64(0)
 	action := NewRuleActionStub(&actionRunCount, &actionCancelCount, t)
 
 	task1 := scheduler.AddTask("Test1")
