+++ nodes/stampzilla-server/logic/task_test.go
@@ -13,8 +13,8 @@ func TestTaskSetUuid(t *testing.T) {
 
 func TestTaskRunAndAddActions(t *testing.T) {
 	task := &task{}
-	actionRunCount := 0
-	actionCancelCount := 0
+	actionRunCount := int64(0)
+	actionCancelCount := int64(0)
 	action := NewRuleActionStub(&actionRunCount, &actionCancelCount, t)
 	task.AddAction(action)
 	task.AddAction(action)
